Added support for baked lighting for Infmaps via vertex color attributes. (also 'fixed' Mee's LUA )
This is great! but not without problems.
For one, the current implementation for rendering infmap worlds does not work with the GShader library.
Another issue is that objects with 'sharp' edges will not look sharp with vertex color lighting. One could potentially fix this issue by implementing face-corner color attributes (along with a custom shader), but that's out of my skillset...
One last thing I'd like to mention is the lack of HDR support for vertex-baked lighting. I imagine this is easily fixed with a custom shader.